Split widgets into separate files
A bit of reorganization. Also some expanded docs, but no behavior or
interface changes yet.
