+++ xi-win-ui/src/widget/mod.rs
@@ -1,0 +1,61 @@
+// Copyright 2018 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Widget trait and common widgets.
+
+use std::any::Any;
+
+use xi_win_shell::paint::PaintCtx;
+use xi_win_shell::window::{MouseButton, MouseType};
+
+use {BoxConstraints, Geometry, LayoutResult};
+use {HandlerCtx, Id, LayoutCtx, PokeCtx};
+
+mod button;
+pub use widget::button::Button;
+
+mod row;
+pub use widget::row::Row;
+
+mod padding;
+pub use widget::padding::Padding;
+
+/// The trait implemented by all widgets.
+pub trait Widget {
+    /// Paint the widget's appearance into the paint context.
+    ///
+    /// The implementer is responsible for translating the coordinates as
+    /// specified in the geometry.
+    #[allow(unused)]
+    fn paint(&mut self, paint_ctx: &mut PaintCtx, geom: &Geometry) {}
+
+    /// Participate in the layout protocol.
+    ///
+    /// `size` is the size of the child previously requested by a RequestChild return.
+    fn layout(&mut self, bc: &BoxConstraints, children: &[Id], size: Option<(f32, f32)>,
+        ctx: &mut LayoutCtx) -> LayoutResult;
+
+    /// Handle a mouse event.
+    ///
+    /// TODO: `MouseType` will be replaced by a click count.
+    #[allow(unused)]
+    fn mouse(&mut self, x: f32, y: f32, mods: u32, which: MouseButton, ty: MouseType,
+        ctx: &mut HandlerCtx) -> bool
+    { false }
+
+    /// An `escape hatch` of sorts for accessing widget state beyond the widget
+    /// methods. Returns true if it is handled.
+    #[allow(unused)]
+    fn poke(&mut self, payload: &mut Any, ctx: &mut PokeCtx) -> bool { false }
+}
